+++ apps/web/src/views/Schedule/MarksByDay/LessonGrades/index.css
@@ -1,0 +1,3 @@
+.flex {
+    gap: 8px;
+}
