+++ apps/web/src/views/Schedule/MarksByDay/index.tsx
@@ -1,43 +1,10 @@
-import { Day } from '@diary-spo/shared'
 import { HorizontalScroll } from '@vkontakte/vkui'
 import { FunctionalComponent } from 'preact'
 import { memo } from 'preact/compat'
+import { IPerformanceCurrent } from './types.ts'
+import { transformData } from './helpers'
+
 import LessonGrades from './LessonGrades'
-import { IPerformanceCurrent, MarkDetailed } from './types.ts'
-
-const transformData = (lessonsState: Day[]): [string, MarkDetailed[]][] => {
-  if (!lessonsState || !lessonsState?.length) {
-    return []
-  }
-
-  const resultMap = new Map<string, MarkDetailed[]>()
-
-  lessonsState.forEach((lessonDay) => {
-    const day = new Date(lessonDay.date).toLocaleDateString('ru')
-
-    lessonDay.lessons?.forEach((lesson) => {
-      if (!lesson.name || !lesson.gradebook) {
-        return
-      }
-
-      const existingTasks = resultMap.get(day) || []
-
-      lesson.gradebook.tasks.forEach((newTask) => {
-        if (
-          !existingTasks.some(
-            (existingTask) => existingTask.task.id === newTask.id
-          )
-        ) {
-          existingTasks.push({ lessonName: lesson.name, task: newTask })
-        }
-      })
-
-      resultMap.set(day, existingTasks)
-    })
-  })
-
-  return Array.from(resultMap.entries())
-}
 
 const MarksByDay: FunctionalComponent<IPerformanceCurrent> = ({
   lessonsState
